+++ qudini-reactive-logging/src/main/java/com/qudini/reactive/logging/web/LoggingContextFilter.java
@@ -1,10 +1,8 @@
 package com.qudini.reactive.logging.web;
 
-import com.qudini.reactive.logging.Log;
 import com.qudini.reactive.logging.ReactiveLoggingContextCreator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.Ordered;
-import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.server.ServerWebExchange;
 import org.springframework.web.server.WebFilter;
 import org.springframework.web.server.WebFilterChain;
@@ -15,29 +13,13 @@ import java.util.Optional;
 import static com.qudini.reactive.utils.MoreTuples.onBoth;
 
 @RequiredArgsConstructor
-public final class LoggingContextFilter implements CorrelationIdForwarder, WebFilter, Ordered {
+public final class LoggingContextFilter implements WebFilter, Ordered {
 
     private final String correlationIdHeader;
 
     private final LoggingContextExtractor loggingContextExtractor;
 
     private final ReactiveLoggingContextCreator reactiveLoggingContextCreator;
-
-    @Override
-    public Mono<WebClient.RequestHeadersSpec<?>> forwardOn(WebClient.RequestHeadersSpec<?> webClient) {
-        return Log
-                .getCorrelationId()
-                .<WebClient.RequestHeadersSpec<?>>map(correlationId -> webClient.header(correlationIdHeader, correlationId))
-                .defaultIfEmpty(webClient);
-    }
-
-    @Override
-    public Mono<WebClient.RequestBodySpec> forwardOn(WebClient.RequestBodySpec webClient) {
-        return Log
-                .getCorrelationId()
-                .map(correlationId -> webClient.header(correlationIdHeader, correlationId))
-                .defaultIfEmpty(webClient);
-    }
 
     @Override
     public Mono<Void> filter(ServerWebExchange exchange, WebFilterChain chain) {
